move test files and fix proptypes
