+++ src/__tests__/ShowOTW.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import ShowOTW from "./ShowOTW.js"; 
+import ShowOTW from "../components/ShowOTW.js"; 
 
 const sampleShow = {
     "title": "Sample Show 1",
