+++ src/components/ShowOTW.js
@@ -33,6 +33,6 @@ ShowOTW.propTypes = {
       duration: PropTypes.number.isRequired,
     }).isRequired,
     genres: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-    id: PropTypes.string.isRequired,
+    id: PropTypes.number.isRequired,
   }).isRequired
 };
