+++ src/components/ShowSnippet.js
@@ -33,6 +33,6 @@ ShowSnippet.propTypes = {
       duration: PropTypes.number.isRequired,
     }).isRequired,
     genres: PropTypes.arrayOf(PropTypes.string.isRequired),
-    id: PropTypes.string.isRequired,
+    id: PropTypes.number.isRequired,
   }).isRequired
 }
